+++ airbyte-integrations/bases/standard-destination-test/src/main/java/io/airbyte/integrations/standardtest/destination/DestinationAcceptanceTest.java
@@ -57,15 +57,20 @@ import java.nio.file.Path;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.HashSet;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Random;
+import java.util.Spliterator;
+import java.util.Spliterators;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import java.util.stream.StreamSupport;
 import org.joda.time.DateTime;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -714,15 +719,21 @@ public abstract class DestinationAcceptanceTest {
 
   @Test
   public void testCustomDbtTransformations() throws Exception {
-    if (!normalizationFromSpec() || !dbtFromSpec()) {
-      // we require normalization implementation for this destination, because we make sure to install
-      // required dbt dependency in the normalization docker image in order to run this test successfully
-      // (we don't actually rely on normalization running anything here though)
+    if (!dbtFromSpec()) {
       return;
     }
 
     final JsonNode config = getConfig();
 
+    // This may throw IllegalStateException "Requesting normalization, but it is not included in the
+    // normalization mappings"
+    // We indeed require normalization implementation of the 'transform_config' function for this
+    // destination,
+    // because we make sure to install required dbt dependency in the normalization docker image in
+    // order to run
+    // this test successfully and that we are able to convert a destination 'config.json' into a dbt
+    // 'profiles.yml'
+    // (we don't actually rely on normalization running anything else here though)
     final DbtTransformationRunner runner = new DbtTransformationRunner(processFactory, NormalizationRunnerFactory.create(
         getImageName(),
         processFactory));
@@ -1016,12 +1027,16 @@ public abstract class DestinationAcceptanceTest {
         .map(AirbyteMessage::getRecord)
         .peek(recordMessage -> recordMessage.setEmittedAt(null))
         .map(recordMessage -> pruneAirbyteInternalFields ? safePrune(recordMessage) : recordMessage)
-        .map(recordMessage -> recordMessage.getData())
+        .map(AirbyteRecordMessage::getData)
+        .peek(this::sortDataFields)
+        .sorted(Comparator.comparing(JsonNode::toString))
         .collect(Collectors.toList());
 
     final List<JsonNode> actualProcessed = actual.stream()
         .map(recordMessage -> pruneAirbyteInternalFields ? safePrune(recordMessage) : recordMessage)
-        .map(recordMessage -> recordMessage.getData())
+        .map(AirbyteRecordMessage::getData)
+        .peek(this::sortDataFields)
+        .sorted(Comparator.comparing(JsonNode::toString))
         .collect(Collectors.toList());
 
     assertSameData(expectedProcessed, actualProcessed);
@@ -1057,6 +1072,21 @@ public abstract class DestinationAcceptanceTest {
         assertSameValue(expectedValue, actualValue);
       }
     }
+  }
+
+  /**
+   * Method that will sort all fields by name and rewrite JsonNode in sorted order
+   *
+   * @param data - data node from AirbyteMessage
+   */
+  private void sortDataFields(JsonNode data) {
+    var sortedFields = StreamSupport.stream(Spliterators.spliteratorUnknownSize(data.fields(),
+        Spliterator.ORDERED), false)
+        .sorted(Entry.comparingByKey(Comparator.comparing(String::toLowerCase)))
+        .collect(Collectors.toList());
+    ((ObjectNode) data).removeAll();
+    IntStream.range(0, sortedFields.size())
+        .forEach(i -> ((ObjectNode) data).set(sortedFields.get(i).getKey(), sortedFields.get(i).getValue()));
   }
 
   // Allows subclasses to implement custom comparison asserts