+++ airbyte-integrations/bases/source-acceptance-test/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
-## 0.1.29
+## 0.1.31
 Take ConfiguredAirbyteCatalog from discover command by default
+
+## 0.1.30
+Validate if each field in a stream has appeared at least once in some record.
+
+## 0.1.29
+Add assert that output catalog does not have any `$ref` in it
 
 ## 0.1.28
 Print stream name when incremental sync tests fail